requirements.txt
Django==2.2.5
django-allauth==0.40.0
django-crispy-forms==1.7.2
django-import-export==1.2.0
djangorestframework==3.10.3
django-environ==0.4.5
psycopg2==2.8.3
Pillow==6.1.0
whitenoise==4.1.3
